Time: 12 ms (13.87%), Space: 56.4 MB (7.28%) - LeetHub

+++ 1354-construct-target-array-with-multiple-sums/1354-construct-target-array-with-multiple-sums.java
@@ -3,7 +3,7 @@ class Solution {
         if (target.length == 1) return target[0] == 1;
 
         PriorityQueue<Integer> pq = new PriorityQueue<>(Collections.reverseOrder());
-        int sum = 0;
+        long sum = 0;
         for (int t: target) {
             pq.add(t);
             sum += t;
@@ -13,7 +13,7 @@ class Solution {
             int curr = pq.poll();
             if (sum - curr == 1) return true;
 
-            int x = curr % (sum - curr);
+            int x = curr % (int)(sum - curr);
             sum = sum - curr + x;
 
             if (x == 0 || x == curr) return false;
